fix: ArchiveCard 정리 및 디자인 수정 (#130)
